+++ src/Configuration.Writable.Abstractions/IReadOnlyNamedOptions.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Microsoft.Extensions.Options;
 
 namespace Configuration.Writable;
 
@@ -13,15 +12,8 @@ public interface IReadOnlyNamedOptions<T> : IReadOnlyOptionsCore<T>
 {
     /// <summary>
     /// Returns a configured <typeparamref name="T"/> instance with the given <paramref name="name"/>.
-    /// This method behaves similarly to the <see cref="IOptionsMonitor{T}.Get(string)"/> method.
     /// </summary>
     T Get(string name);
-
-    /// <summary>
-    /// Retrieves the configuration settings object for the specified configuration section name.
-    /// </summary>
-    /// <param name="name">The name of the configuration section to retrieve options for.</param>
-    WritableOptionsConfiguration<T> GetOptionsConfiguration(string name);
 
     /// <summary>
     /// Retrieves a read-only options instance bound to the specified instanceName.
@@ -33,9 +25,7 @@ public interface IReadOnlyNamedOptions<T> : IReadOnlyOptionsCore<T>
     IReadOnlyOptions<T> GetInstance(string name);
 
     /// <summary>
-    /// Registers a listener to be called whenever a named <typeparamref name="T"/> changes. <br/>
-    /// This method behaves similarly to the <see cref="IOptionsMonitor{T}.OnChange"/> method, <br/>
-    /// but is only called for the matching instanceName.
+    /// Registers a listener to be called whenever the matching named <typeparamref name="T"/> changes.
     /// </summary>
     /// <param name="name">The name of the options instance to listen for changes.</param>
     /// <param name="listener">The action to be invoked when <typeparamref name="T"/> has changed.</param>