+++ src/Configuration.Writable.Abstractions/IReadOnlyOptionsCore.cs
@@ -1,6 +1,5 @@
 using System;
 using System.ComponentModel;
-using Microsoft.Extensions.Options;
 
 namespace Configuration.Writable;
 
@@ -12,8 +11,7 @@ public interface IReadOnlyOptionsCore<T>
     where T : class, new()
 {
     /// <summary>
-    /// Registers a listener to be called whenever a named <typeparamref name="T"/> changes. <br/>
-    /// This method behaves similarly to the <see cref="IOptionsMonitor{T}.OnChange"/> method.
+    /// Registers a listener to be called whenever a named <typeparamref name="T"/> changes.
     /// </summary>
     /// <param name="listener">The action to be invoked when <typeparamref name="T"/> has changed.</param>
     /// <returns>An <see cref="IDisposable"/> which should be disposed to stop listening for changes.</returns>