+++ tests/Configuration.Writable.Tests/WritableOptionsExtensionsTests.cs
@@ -87,7 +87,7 @@ public partial class WritableOptionsExtensionsTests
         var readonlyOptions = serviceProvider.GetService<IReadOnlyOptions<TestSettings>>();
         writableOptions.ShouldNotBeNull();
         readonlyOptions.ShouldNotBeNull();
-        writableOptions.GetOptionsConfiguration().ConfigFilePath.ShouldBe(testFilePath);
+        writableOptions.ConfigurationInfo.WritePath.ShouldBe(testFilePath);
     }
 
     [Fact]
@@ -119,8 +119,7 @@ public partial class WritableOptionsExtensionsTests
         var host = builder.Build();
         var writableOptions = host.Services.GetRequiredService<IWritableOptions<TestSettings>>();
 
-        var configOptions = writableOptions.GetOptionsConfiguration();
-        configOptions.ConfigFilePath.ShouldBe(testFilePath);
+        writableOptions.ConfigurationInfo.WritePath.ShouldBe(testFilePath);
     }
 
     [Fact]